+++ src/components/RichTexts/index.tsx
@@ -1,0 +1,13 @@
+export * from "./Annotations";
+export * from "./Bold";
+export * from "./Color";
+export * from "./Equation";
+export * from "./InlineCode";
+export * from "./Italic";
+export * from "./Link";
+export * from "./Mention";
+export * from "./RichText";
+export * from "./RichTexts";
+export * from "./Strikethrough";
+export * from "./Text";
+export * from "./Underline";
